requirements/requirements.txt
wheel==0.32.2

# Python-PostgreSQL Database Adapter
psycopg2==2.7.6

# Django
django==2.1.3
django-environ==0.4.5
django-filter==2.0.0
djangorestframework==3.9.0
django-extensions==2.1.3

# Time zones support
pytz==2018.5

# runserver_plus support
Werkzeug==0.12

# Password storage
argon2-cffi==18.1.0






django-rest-swagger~=2.2

git+git://github.com/philipn/django-rest-framework-filters@master


django-braces~=1.13
django-crispy-forms==1.7.2

django-model-utils~=3.1

# Images
Pillow~=5.2


# Unicode slugification
awesome-slugify==1.6.5



# Your custom requirements go here

django-storages>=1.6



# Admin site
django-admin-json-editor>=0.1.5
django-import-export~=1.0
django-map-widgets~=0.2
